Вынесет константы
